feat: 다이어리 배경지 부분 컴포넌트 분리
